chore: clear the compiler and lint warnings 🤖
Kotlin warnings (8, all pre-existing):

- `build.gradle.kts`: AGP 9 deprecates srcDir/srcDirs/setSrcDirs on
  AndroidSourceDirectorySet; use the `directories` set instead. The schema export
  dir the androidTest MigrationTestHelper reads still lands in the test APK
  (verified against app-debug-androidTest.apk).
- `DictationScreen`: MIN_INTERVAL_SEC/MAX_INTERVAL_SEC are already Double — drop
  the redundant toDouble(), matching HomePlaybackPanel's form.
- `LibraryViewModel`: debounce() needs FlowPreview; opt in next to the existing
  ExperimentalCoroutinesApi, as HomeViewModel already does.
- `YoudaoTts` / `SystemSpeaker`: unnecessary safe calls (okhttp 5's Response.body
  is non-null; the holder is assigned unconditionally before that branch). The
  Youdao read stays wrapped in its catch-to-null boundary.
- `OcrScanSheet`: deprecated TabRow → PrimaryTabRow.
- `SettingsProviderPages`: deprecated MenuAnchorType typealias →
  ExposedDropdownMenuAnchorType (4 call sites).
- `io/edge/EdgeTts.kt`: code untouched — the file is vendored and documents
  compiling against okhttp 4.12, where Response.body is nullable; removing the
  `?.` would break that portability contract. UNNECESSARY_SAFE_CALL is folded
  into the existing @file:Suppress (the annotation is not repeatable) with a note
  on the version difference.

Lint warnings (3, found by lintDebug):

- `TtsTestStatusLine`: a default of `Modifier.padding(...)` violates
  ModifierParameter. No call site passed it, so the parameter is gone and the
  padding is now a local modifier — same rendering.
- `HearWriteWordmark`: move `modifier` ahead of `tileSize`; both call sites use
  named arguments.
- `HomeScreen`: mutableStateOf(0) → mutableIntStateOf(0), dropping the Int boxing.

Verified with `assembleDebug assembleDebugAndroidTest testDebugUnitTest
lintDebug --rerun-tasks` (full recompile, not incremental): zero `w:`/`e:` lines,
289 unit tests green, lint 0 findings. Emulator pixel_9a_api37 walkthrough of the
four touched surfaces: OCR sheet tab indicator moves 英文→中文 on tap, the 试听
status line keeps its spinner+text row and the sections below shift correctly,
the About wordmark stays centered at 44.dp, the dictation panel steps 7.0s↔8.0s
and clamps at 1.0s, and the library search renders after the debounce.

+++ app/src/main/java/org/yangtse/hearwrite/ui/DictationScreen.kt
@@ -686,10 +686,10 @@ private fun PlaybackPanel(
                 Spacer(Modifier.width(8.dp))
                 IconButton(
                     onClick = {
-                        val stepped = ((ui.intervalSec - 0.5).coerceIn(
-                            MIN_INTERVAL_SEC.toDouble(),
-                            MAX_INTERVAL_SEC.toDouble(),
-                        ))
+                        val stepped = (ui.intervalSec - 0.5).coerceIn(
+                            MIN_INTERVAL_SEC,
+                            MAX_INTERVAL_SEC,
+                        )
                         onIntervalChange(stepped.toFloat())
                     },
                     enabled = ui.intervalSec > MIN_INTERVAL_SEC,
@@ -705,10 +705,10 @@ private fun PlaybackPanel(
                 )
                 IconButton(
                     onClick = {
-                        val stepped = ((ui.intervalSec + 0.5).coerceIn(
-                            MIN_INTERVAL_SEC.toDouble(),
-                            MAX_INTERVAL_SEC.toDouble(),
-                        ))
+                        val stepped = (ui.intervalSec + 0.5).coerceIn(
+                            MIN_INTERVAL_SEC,
+                            MAX_INTERVAL_SEC,
+                        )
                         onIntervalChange(stepped.toFloat())
                     },
                     enabled = ui.intervalSec < MAX_INTERVAL_SEC,
